Fix "yydebug undeclared" error
YYDEBUG is always defined in y.tab.c (hoc.c):

    #ifndef YYDEBUG
    # define YYDEBUG 0
    #endif

so in hoc.y, using:

    #ifdef YYDEBUG

is incorrect. It should be:

    #if YYDEBUG

as used elsewhere in y.tab.c (hoc.c).

Also yydebug is defined in y.tab.c (hoc.c), so the extern declaration of it in
hoc.y is not needed.